Improved views in the ServicesAdapter
App now can dynamically change the service information upon the user's role and status.
Code cleaned and commented in some sections.
App now uploads (to a maximum of 5) the number of pics of carers.
App now deletes the image that is shown in the PagerAdapter (WIP).
Service layout is divided depending upon the service status.
Review CRUD check and adding methods in Firebase implemented (WIP).