Added Reviews SheetDialog
Owner now can add reviews on completed services.
App checks if there is a review for that service (only one review per service).
App inserts in Firebase the review (subdocument in the Carer's document).
App updates the Carer's rating with a 'SELECT + UPDATE' operation in Firebase.
Designed SheetDialog for reviewing services.
Designed and implemented the Adapter and Layout for the Reviews RecyclerView.